Improvised the implementation

+++ utils/github.py
@@ -3,6 +3,7 @@ import json
 import sys
 import time
 import urllib2
+import os
 from getpass import getpass
 
 from utils.cmd import keep_trying
@@ -121,20 +122,38 @@ def github_list_pull_requests(urls, numbers_only=False):
         created_at = time.strptime(created_at, "%Y-%m-%dT%H:%M:%SZ")
         created_at = time.mktime(created_at)
         user = pull["user"]["login"]
-        with open('author_info') as infile:
-            data = json.load(infile)
-        got_it = False
-        for k in iter(data):
-            if k == user:
-                author = "\"%s\" <%s>" % (data[user]["name"], data[user]["email"])
-                got_it = True
-        if not got_it:
+        auth_dir = os.path.dirname("~/.sympy/author_info") # Check for the existence of .sympy directory
+        auth_dir = os.path.expanduser(auth_dir)
+        if not os.path.exists(auth_dir):
+            # For users who have not yet configured sympy-bot
             user_info = github_get_user_info(urls, user)
             author = "\"%s\" <%s>" % (user_info.get("name", "unknown"),
                                       user_info.get("email", ""))
-            data[user_info["login"]] = {"name": user_info.get("name", "unknown"), "email": user_info.get("email", "")}
-            with open('author_info', 'wb') as outfile:
-                json.dump(data, outfile, indent=2)
+
+        else:
+            auth_file = os.path.normpath("~/.sympy/author_info") #If it exists check for author_info file
+            auth_file = os.path.expanduser(auth_file)
+
+            if not os.path.exists(auth_file):      #If it doesn't exist, then create an empty file
+                data = dict()
+                with open(auth_file, 'wb') as outfile:
+                    json.dump(data, outfile, indent=2)
+
+            with open(auth_file) as infile:
+                data = json.load(infile)
+            got_it = False
+            for k in iter(data):
+                if k == user:
+                    author = "\"%s\" <%s>" % (data[user]["name"], data[user]["email"])
+                    got_it = True
+            if not got_it:
+                user_info = github_get_user_info(urls, user)
+                author = "\"%s\" <%s>" % (user_info.get("name", "unknown"),
+                                          user_info.get("email", ""))
+                data[user_info["login"]] = {"name": user_info.get("name", "unknown"), "email": user_info.get("email", "")}
+                with open(auth_file, 'wb') as outfile:
+                    json.dump(data, outfile, indent=2)
+
         branch_against = pull["base"]["ref"]
         formatted_pulls.append({
             'created_at': created_at,
